Refactor p1 and split visitor and structural checks modules
